refactor: Streamline PDF file processing by introducing usePdfTasks hook for improved state management and task polling
